nits + Move updatePlanElement to deploy.ts instead of api.ts
